feat: simplify & redesign the create guild page
