+++ src/pages/create-guild.tsx
@@ -1,34 +1,14 @@
-import { ThemeProvider, useThemeContext } from "components/[guild]/ThemeContext"
+import { Box, Heading, Stack, useColorModeValue } from "@chakra-ui/react"
+import { ThemeProvider } from "components/[guild]/ThemeContext"
 import ClientOnly from "components/common/ClientOnly"
 import { Layout } from "components/common/Layout"
+import Head from "components/common/Layout/components/Head"
 import CreateGuildForm, {
   CreateGuildFormType,
 } from "components/create-guild/CreateGuildForm"
 import DynamicDevTool from "components/create-guild/DynamicDevTool"
-import { FormProvider, useForm, useFormContext, useWatch } from "react-hook-form"
+import { FormProvider, useForm } from "react-hook-form"
 import getRandomInt from "utils/getRandomInt"
-
-function CreateGuildHead() {
-  const name = useWatch({ name: "name" })
-  const imageUrl = useWatch({ name: "imageUrl" })
-
-  return <Layout.Head ogTitle={name || "Create Guild"} imageUrl={imageUrl} />
-}
-
-function CreateGuildBackground() {
-  const { localThemeColor, localBackgroundImage } = useThemeContext()
-  const themeColor = useWatch({ name: "theme.color" })
-  const color = localThemeColor !== themeColor ? themeColor : localThemeColor
-
-  return (
-    <Layout.Background offset={47} background={color} image={localBackgroundImage} />
-  )
-}
-
-function CreateGuildDynamicDevTool() {
-  const { control } = useFormContext<CreateGuildFormType>()
-  return <DynamicDevTool control={control} />
-}
 
 const CreateGuildPage = (): JSX.Element => {
   const methods = useForm<CreateGuildFormType>({
@@ -45,24 +25,59 @@ const CreateGuildPage = (): JSX.Element => {
     },
   })
 
-  return (
-    <ClientOnly>
-      <FormProvider {...methods}>
-        <Layout.Root maxWidth="sizes.xl">
-          <CreateGuildHead />
-          <Layout.HeaderSection>
-            <CreateGuildBackground />
-            <Layout.Header />
-            <Layout.Headline title="Begin your guild" />
-          </Layout.HeaderSection>
+  const bgColor = useColorModeValue("white", "var(--chakra-colors-gray-700)")
+  const bgFile = useColorModeValue("bg_light.svg", "bg.svg")
 
-          <Layout.MainSection>
-            <CreateGuildForm />
-          </Layout.MainSection>
-        </Layout.Root>
-        <CreateGuildDynamicDevTool />
-      </FormProvider>
-    </ClientOnly>
+  return (
+    <>
+      <Head ogTitle="Begin your guild" />
+      <ClientOnly>
+        <FormProvider {...methods}>
+          <Box
+            display="grid"
+            gridTemplateRows="auto,1fr"
+            gap={8}
+            h="100vh"
+            position="relative"
+            _before={{
+              content: '""',
+              position: "absolute",
+              top: 0,
+              bottom: 0,
+              left: 0,
+              right: 0,
+              bg: `radial-gradient(circle at 50% 75%, ${bgColor} 60%, transparent), url('/landing/${bgFile}')`,
+              bgSize: "100% auto",
+              bgPosition: "top 0.5rem center",
+              opacity: "0.07",
+              zIndex: -1,
+            }}
+          >
+            <Layout.Header />
+
+            <Stack
+              spacing={8}
+              pb={8}
+              px={4}
+              w="min(100%, var(--chakra-sizes-md))"
+              mx="auto"
+            >
+              <Heading
+                as="h2"
+                fontFamily="display"
+                textAlign="center"
+                fontSize={{ base: "4xl", sm: "5xl" }}
+              >
+                Begin your guild
+              </Heading>
+              <CreateGuildForm />
+            </Stack>
+          </Box>
+
+          <DynamicDevTool control={methods.control} />
+        </FormProvider>
+      </ClientOnly>
+    </>
   )
 }
